refactor(components): split dropdown in small components
